requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
astropy==6.0.0
astropy-iers-data==0.2024.1.8.0.30.55
attrs==23.2.0
audioread==3.0.1
biopython==1.84
blinker==1.9.0
cachetools==5.5.0
caer==2.0.8
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
ete3==3.1.3
filelock==3.12.3
Flask==3.1.0
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.9.2
gensim==4.3.2
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
hmmlearn==0.3.2
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.4
imbalanced-learn==0.12.4
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.8.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy==1.5.1
mypy-extensions==1.0.0
networkx==3.1
nltk==3.9.1
numba==0.59.1
numpy==1.26.1
openai==0.28.0
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.68
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
platformdirs==4.2.2
pooch==1.8.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyerfa==2.0.1.1
pygame==2.5.2
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.11.6
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.3
seaborn==0.13.0
selenium==4.22.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
tensorflow-addons==0.22.0
threadpoolctl==3.2.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.67.1
trio==0.25.1
trio-websocket==0.11.1
twilio==9.2.2
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
